Remake ResourceLink
