+++ src/EduManagementLab.Web/Pages/CourseTasks/CourseTaskResourceLinks.cshtml.cs
@@ -4,6 +4,7 @@ using EduManagementLab.Core.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using System.ComponentModel.DataAnnotations;
 
 namespace EduManagementLab.Web.Pages.CourseLineItems
 {
@@ -12,13 +13,26 @@ namespace EduManagementLab.Web.Pages.CourseLineItems
     {
         private readonly CourseTaskService _courseTaskService;
         private readonly ResourceLinkService _resourceLinkService;
-        public CourseTaskResourceLinksModel(CourseTaskService courseLineItemService, ResourceLinkService resourceLinkService)
+        private readonly ToolService _toolService;
+        public CourseTaskResourceLinksModel(CourseTaskService courseLineItemService, ResourceLinkService resourceLinkService, ToolService toolService)
         {
             _courseTaskService = courseLineItemService;
             _resourceLinkService = resourceLinkService;
+            _toolService = toolService;
         }
         public List<SelectListItem> ResourceLinkSelectList { get; } = new List<SelectListItem>();
-        public string SelctedResource { get; set; }
+        [BindProperty]
+        public ResourceLinkInputModel ResourceLink { get; set; } = new ResourceLinkInputModel();
+        public class ResourceLinkInputModel
+        {
+            public string? CustomProperties { get; set; }
+            public string? Description { get; set; }
+            [Required]
+            public string Title { get; set; }
+            [Required]
+            public Guid selectedTool { get; set; }
+        }
+        public List<SelectListItem> tools { get; set; } = new List<SelectListItem>();
         public CourseTask courseTask { get; set; }
         public IActionResult OnGet(Guid courseTaskId)
         {
@@ -42,41 +56,51 @@ namespace EduManagementLab.Web.Pages.CourseLineItems
                 ResourceLinkSelectList.Add(new SelectListItem { Text = resourceLink.Title, Value = resourceLink.Id.ToString() });
             }
 
-            //foreach (var resourceLink in _resourceLinkService.GetResourceLinks()
-            //    .Where(s => !courseTask.IMSLTIResourceLinks.Any(x => x.Id == s.Id) &&
-            //    courseTaskList.Any(c => !c.IMSLTIResourceLinks.Any(r => r.Id == s.Id))))
-            //{
-
-            //    ResourceLinkSelectList.Add(new SelectListItem { Text = resourceLink.Title, Value = resourceLink.Id.ToString() });
-            //}
+            foreach (var tool in _toolService.GetTools())
+            {
+                tools.Add(new SelectListItem { Text = tool.Name, Value = tool.Id.ToString() });
+            }
         }
-        public IActionResult OnPostAddResourceLinkToCourseTask(Guid courseTaskId)
+        public IActionResult OnPost(Guid courseTaskId)
         {
-            try
-            {
-                _courseTaskService.AddResouceLinkToCourseTask(courseTaskId, Guid.Parse(SelctedResource));
-            }
-            catch (CourseTaskNotFoundException ex)
-            {
-                return NotFound(ex.Message);
-            }
+            var targetTool = _toolService.GetTool(ResourceLink.selectedTool);
 
-            LoadData(courseTaskId);
-            return Page();
-        }
-        public IActionResult OnPostRemoveResourceLinkFromCourseTask(Guid resourceId, Guid courseTaskId)
-        {
-            try
+            if (targetTool != null)
             {
-                _courseTaskService.DeleteCourseTaskResoruceLink(courseTaskId, resourceId);
-            }
-            catch (CourseTaskNotFoundException ex)
-            {
-                return NotFound(ex.Message);
+                IMSLTIResourceLink newResource = new IMSLTIResourceLink()
+                {
+                    Description = ResourceLink.Description,
+                    Title = ResourceLink.Title,
+                    Tool = targetTool,
+                    CustomProperties = ResourceLink.CustomProperties,
+                };
+
+                try
+                {
+                    _resourceLinkService.CreateResourceLink(newResource);
+                    _courseTaskService.AddResouceLinkToCourseTask(courseTaskId, newResource);
+                }
+                catch (CourseTaskNotFoundException ex)
+                {
+                    return NotFound(ex.Message);
+                }
             }
 
             LoadData(courseTaskId);
             return RedirectToPage("./CourseTaskResourceLinks", new { courseTaskId = courseTaskId });
+        }
+        public IActionResult OnPostRemoveResourceLink(Guid resourceId, Guid courseTaskId)
+        {
+            try
+            {
+                _resourceLinkService.DeleteResourceLink(resourceId);
+            }
+            catch (CourseTaskNotFoundException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            LoadData(courseTaskId);
+            return Page();
         }
     }
 }
