+++ src/EduManagementLab.Web/Pages/ResourceLinks/Edit.cshtml
@@ -5,7 +5,7 @@
 <div class="container-fluid">
     <h4>Add ResourceLink</h4>
     <hr />
-    <form method="post" asp-route-id="@Model.ResourceLink.Id">
+    <form method="post" asp-route-id="@Model.ResourceLink.Id" asp-route-CourseTaskId="@Model.CourseTaskId">
         <div asp-validation-summary="ModelOnly" class="text-danger"></div>
         <div class="row">
             <div class="col-lg-6 mb-3">
