--- src/EduManagementLab.Web/Pages/ResourceLinks/Index.cshtml
@@ -1,37 +1,0 @@
-﻿@page
-@model EduManagementLab.Web.Pages.ResourceLinks.IndexModel
-@{
-}
-
-<a class="btn btn-primary m-2" asp-page="/ResourceLinks/CreateResourceLink">Add ResourceLink</a>
-@if (@Model.ResourceLinksList.Any())
-{
-    <div class="row">
-        @foreach (var resource in @Model.ResourceLinksList)
-        {
-            <div class="col-sm-3">
-                <div class="card">
-                    <div class="card-body">
-                        <h5 class="card-title"><strong>Title: </strong>@resource.Title</h5>
-                        <p class="card-text"><strong>Description: </strong> @resource.Description</p>
-                        <p class="card-text"><strong>Tool Launch URL: </strong>@resource.Tool.LaunchUrl</p>
-
-                        <hr>
-                        <a asp-page="Detail" asp-route-resourceId="@resource.Id" class="btn btn-primary m-2">Detail</a>
-                        <a asp-page="Edit" asp-route-resourceId="@resource.Id"class="btn btn-primary m-2">Edit</a>
-                        <form method="post" asp-page-handler="DeleteResourceLink" asp-route-resourceId="@resource.Id">
-                            <button type="submit" class="btn btn-danger m-2">Delete</button>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        }
-    </div>
-}
-else
-{
-    <div class="position-relative">
-        <h1 class="text-center">No ResourceLinks available</h1>
-    </div>
-}
-
