+++ src/EduManagementLab.Web/Pages/Shared/_Layout.cshtml
@@ -31,9 +31,9 @@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-page="/Tools/Index">Tools</a>
                         </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-page="/ResourceLinks/Index">ResourceLinks</a>
-                        </li>
+@*                        <li class="nav-item">
+                                <a class="nav-link text-dark" asp-area="" asp-page="/ResourceLinks/Index">ResourceLinks</a>
+                            </li>*@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
                         </li>
@@ -60,12 +60,12 @@
     <script src="~/js/site.js" asp-append-version="true"></script>
 
     @*@using IdentityServer4.Extensions
-        @inject EduManagementLab.Core.Services.CourseService courseService;
-        @inject EduManagementLab.Core.Services.UserService userService;
-        @{
-        var userId = Guid.Parse(Context.User?.GetSubjectId());
-        var user = userService.GetUser(userId);
-        }*@
+            @inject EduManagementLab.Core.Services.CourseService courseService;
+            @inject EduManagementLab.Core.Services.UserService userService;
+            @{
+            var userId = Guid.Parse(Context.User?.GetSubjectId());
+            var user = userService.GetUser(userId);
+            }*@
     @await RenderSectionAsync("Scripts", required: false)
 </body>
 </html>
